requirements.txt
Django==1.7.4
Django-Select2==4.2.2
Pillow==2.7.0
argparse==1.2.1
beautifulsoup4==4.3.2
django-braces==1.4.0
django-ckeditor==4.4.7
django-debug-toolbar==1.2.2
django-filter==0.9.2
django-htmlmin==0.7.0
djangorestframework==3.0.4
easy-thumbnails==2.2
gunicorn==19.2.0
html5lib==0.999
ipython==2.4.0
psycopg2==2.5.4
pylibmc==1.4.1
pytils==0.3
setproctitle==1.1.8
six==1.9.0
sqlparse==0.1.14
wsgiref==0.1.2
